+++ tests/bootstrap.php
@@ -5,6 +5,9 @@ if ( ! $_tests_dir ) {
 	$_tests_dir = '/tmp/wordpress-tests-lib';
 }
 
+// Load the composer autoloader.
+require_once __DIR__ . '/../vendor/autoload.php';
+
 require_once $_tests_dir . '/includes/functions.php';
 
 function _manually_load_plugin() {