+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog for Co-Authors Plus
 
+**3.5.6 (Nov 16, 2022)**
+* CI: Add new deployment workflow #890
+* Dependency update #891
+
+**3.5.5 (Nov 16, 2022)**
+* Add Yoast integration #877
+* CI: Add WP 6.1 #886 #887
+
+**3.5.4 (Nov 4, 2022)**
+* Handle `WP_Error` on failed guest author creation for method `create_guest_author()` #879
+* Only use `pre_handle_404` filter in non-paged result #874
+* Add CLI create-author command #880
+* Add template tag `co_authors_get_users()` #862
+* Change permission callback for authors and search endpoint and improve `current_user_can_set_authors()` #883
+* Fix tests and add IDE files #882
+* Bump dependencies #866 #865 #864
+
+**3.5.3 (Oct 24, 2022)**
+* Add author taxonomy labels #860
+* Add check for empty author in Jetpack Open Graph tags #861
+* Improve performance of search_authors() #872
+* Update WP versions in tests and grant permission to composer-installer packages #863
+
 **3.5.2 (June 1, 2022)**
 * Remove the `/wp/v2/coauthors` endpoint #851
 * Remove CSS hack that hides the add author box from core in the block editor #847